requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
datasets>=2.18.0
lpips>=0.1.4
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
